Move FileSaveWatcher to its own file
